expand module rewards
