+++ pallets/parachain-staking/src/rewards/transfer_from_rewards_account.rs
@@ -1,0 +1,1 @@
+
